refactor(desktop): retire the Fleet page into Machine
